Split messages files and clang format
